Added getting data source for reports creating

+++ src/main/kotlin/digital/tutors/autochecker/reviewer/entities/PeerReview.kt
@@ -24,7 +24,7 @@ data class PeerReview(
         var solutionId: PeerTaskSolution? = null,
 
         var grade: Double = 0.0,
-        var gradesPerCriterions: List<Int>? = null,
+        var gradesPerCriterions: List<Double>? = null,
         var argumentsPerCriterions: List<String>? = null,
         var summaryMessagePerSolution: String? = null
 ) : AuditableEntity()
