+++ src/main/kotlin/digital/tutors/autochecker/reviewer/vo/peerReview/PeerReviewUpdateRq.kt
@@ -3,7 +3,7 @@ package digital.tutors.autochecker.reviewer.vo.peerReview
 import java.util.*
 
 data class PeerReviewUpdateRq(
-        var gradesPerCriterions: List<Int>?,
+        var gradesPerCriterions: List<Double>?,
         var argumentsPerCriterions: List<String>?,
         var summaryMessagePerSolution: String?,
         var grade: Double?
